Update }bedrock.cube.data.export.pro
Existing public views should not be recreated. The use case is that views already exist (typically with filters applied) and then a simple call to '}bedrock.cube.data.export' with the name of the view and pFilter = '' would reset the view and remove the filters. That reset is done in '}bedrock.cube.view.create' but probably best to not break into that process.